+++ src/Order/Repository/OrderRepositoryInterface.php
@@ -20,6 +20,4 @@ interface OrderRepositoryInterface
     public function getByOrderId(string $orderId): OrderModel;
 
     public function fillEmptyInvoiceNumber(OrderModel $orderModel): void;
-
-    public function getInvoiceNumberByOrderId(string $orderId): string;
 }